+++ zebra-chain/src/transaction/tests/vectors.rs
@@ -1,0 +1,65 @@
+use super::super::*;
+
+use crate::serialization::{ZcashDeserialize, ZcashSerialize};
+
+#[test]
+fn librustzcash_tx_deserialize_and_round_trip() {
+    let tx = Transaction::zcash_deserialize(&zebra_test::vectors::GENERIC_TESTNET_TX[..])
+        .expect("transaction test vector from librustzcash should deserialize");
+
+    let mut data2 = Vec::new();
+    tx.zcash_serialize(&mut data2).expect("tx should serialize");
+
+    assert_eq!(&zebra_test::vectors::GENERIC_TESTNET_TX[..], &data2[..]);
+}
+
+#[test]
+fn zip143_deserialize_and_round_trip() {
+    let tx1 = Transaction::zcash_deserialize(&zebra_test::vectors::ZIP143_1[..])
+        .expect("transaction test vector from ZIP143 should deserialize");
+
+    let mut data1 = Vec::new();
+    tx1.zcash_serialize(&mut data1)
+        .expect("tx should serialize");
+
+    assert_eq!(&zebra_test::vectors::ZIP143_1[..], &data1[..]);
+
+    let tx2 = Transaction::zcash_deserialize(&zebra_test::vectors::ZIP143_2[..])
+        .expect("transaction test vector from ZIP143 should deserialize");
+
+    let mut data2 = Vec::new();
+    tx2.zcash_serialize(&mut data2)
+        .expect("tx should serialize");
+
+    assert_eq!(&zebra_test::vectors::ZIP143_2[..], &data2[..]);
+}
+
+#[test]
+fn zip243_deserialize_and_round_trip() {
+    let tx1 = Transaction::zcash_deserialize(&zebra_test::vectors::ZIP243_1[..])
+        .expect("transaction test vector from ZIP243 should deserialize");
+
+    let mut data1 = Vec::new();
+    tx1.zcash_serialize(&mut data1)
+        .expect("tx should serialize");
+
+    assert_eq!(&zebra_test::vectors::ZIP243_1[..], &data1[..]);
+
+    let tx2 = Transaction::zcash_deserialize(&zebra_test::vectors::ZIP243_2[..])
+        .expect("transaction test vector from ZIP243 should deserialize");
+
+    let mut data2 = Vec::new();
+    tx2.zcash_serialize(&mut data2)
+        .expect("tx should serialize");
+
+    assert_eq!(&zebra_test::vectors::ZIP243_2[..], &data2[..]);
+
+    let tx3 = Transaction::zcash_deserialize(&zebra_test::vectors::ZIP243_3[..])
+        .expect("transaction test vector from ZIP243 should deserialize");
+
+    let mut data3 = Vec::new();
+    tx3.zcash_serialize(&mut data3)
+        .expect("tx should serialize");
+
+    assert_eq!(&zebra_test::vectors::ZIP243_3[..], &data3[..]);
+}
